+++ frontend/src/pages/games/hooks/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Games hooks exports
+ */
+
+export { useGameScores } from './useGameScores'
+export { useGameTimer } from './useGameTimer'
+export { useKeyboard } from './useKeyboard'