+++ frontend/src/pages/games/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Games module exports
+ */
+
+export * from './types'
+export * from './constants'